rename catalog-get-tokenuri to zora-get-tokenuri
