+++ test/strategies/zora-get-tokenuri/transformer_test.mjs
@@ -5,7 +5,7 @@ import addFormats from "ajv-formats";
 
 import { track } from "@neume-network/schema";
 
-import { onLine } from "../../../src/strategies/catalog-get-tokenuri/transformer.mjs";
+import { onLine } from "../../../src/strategies/zora-get-tokenuri/transformer.mjs";
 
 const ajv = new Ajv();
 addFormats(ajv);
@@ -52,7 +52,7 @@ test("catalog transformer", (t) => {
   const valid = validate(data);
   t.is(validate.errors.length, 1);
   t.is(data.erc721.tokenURI, payload.metadata.tokenURI);
-  // NOTE: catalog-get-tokenuri implements neume-network/schema only partially
+  // NOTE: zora-get-tokenuri implements neume-network/schema only partially
   // at this point, so we can't expect it to pass validation. Instead, we're
   // expecting errors to be thrown.
   t.is(validate.errors[0].params.missingProperty, "address");
